myPage reservationCode 타입 수정 (#109)

+++ metanet4/src/main/java/com/metanet/team4/mypage/model/ReserveList.java
@@ -17,7 +17,7 @@ public class ReserveList {
 	private String cinemaName;
 	private String screenName;
 	private String seatName;
-	private int reservationCode;
+	private Long reservationCode;
 	private Date reservationTime;
 	private int paymentAmount;
 }
